+++ Wikipedia/assets/codemirror/codemirror-index.html
@@ -38,7 +38,9 @@
 
 
 RangeHelper.rangeDebugging.showRangeDebuggingButtons()
-//const markupItems = RangeHelper.rangeDetermination.markupItemsForLine(1)
+//const markupItems = RangeHelper.rangeDetermination.markupItemsForLineTokens(editor.getLineTokens(1, true))
+
+
