+++ www/codemirror/codemirror-range-determination-non-tag.js
@@ -1,0 +1,70 @@
+const intersection = require('./codemirror-set-utilities').intersection
+const difference = require('./codemirror-set-utilities').difference
+
+const ItemRange = require('./codemirror-range-objects').ItemRange
+const MarkupItem = require('./codemirror-range-objects').MarkupItem
+
+// - returns set of types for token
+// - smooths out inconsistent nested bold and italic types
+const tokenTypes = (token) => {
+  const types = (token.type || '')
+    .trim()
+    .split(' ')
+    .filter(s => s.length > 0)
+    .map(s => {
+      // the parser fails to add 'mw-apostrophes-bold' for nested bold (it only adds 'mw-apostrophes')
+      if (s === 'mw-apostrophes' && token.string === `'''`) {
+        return 'mw-apostrophes-bold'
+      }
+      // the parser fails to add 'mw-apostrophes-italic' for nested italic (it only adds 'mw-apostrophes')
+      if (s === 'mw-apostrophes' && token.string === `''`) {
+        return 'mw-apostrophes-italic'
+      }
+      return s
+    })
+
+  return new Set(types)
+}
+
+const nonTagMarkupItemsForLineTokens = (lineTokens) => {
+  const soughtTokenTypes = new Set(['mw-apostrophes-bold', 'mw-apostrophes-italic', 'mw-link-bracket', 'mw-section-header', 'mw-template-bracket'])  
+
+  let trackedTypes = new Set()
+  let outputMarkupItems = []
+  
+  const tokenWithEnrichedInHtmlTagArray = (token, index, tokens) => {
+    
+    const types = intersection(tokenTypes(token), soughtTokenTypes)
+    
+    const typesToStopTracking = Array.from(intersection(trackedTypes, types))
+    const typesToStartTracking = Array.from(difference(types, trackedTypes))
+    
+    const addMarkupItemWithRangeStarts = (type) => {
+      const inner = new ItemRange(token.end, -1) 
+      const outer = new ItemRange(token.start, -1) 
+      const markupItem = new MarkupItem(type, inner, outer)
+      outputMarkupItems.push(markupItem)
+    }
+    
+    const updateMarkupItemRangeEnds = (type) => {
+      const markupItem = outputMarkupItems.find(markupItem => {
+        return markupItem.type === type && !markupItem.isComplete()
+      })
+      if (markupItem) {
+        markupItem.inner.end = token.start
+        markupItem.outer.end = token.end
+      }
+    }
+    
+    typesToStartTracking.forEach(addMarkupItemWithRangeStarts)
+    typesToStopTracking.forEach(updateMarkupItemRangeEnds)
+    
+    typesToStopTracking.forEach(tag => trackedTypes.delete(tag))
+    typesToStartTracking.forEach(tag => trackedTypes.add(tag))
+  }
+  
+  lineTokens.forEach(tokenWithEnrichedInHtmlTagArray)    
+  return outputMarkupItems
+}
+
+exports.nonTagMarkupItemsForLineTokens = nonTagMarkupItemsForLineTokens
